pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/docgemma"]

[project]
name = "docgemma-connect"
version = "0.1.0"
description = "Agentic medical AI with autonomous tool calling, powered by MedGemma 4B for resource-limited healthcare environments"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.12.0",
    "httpx>=0.28.0",
    "huggingface-hub>=1.3.4",
    "langgraph>=0.2.0",
    "mcp>=1.0.0",
    "outlines>=1.2.9",
    "pillow>=12.1.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.2.1",
    "transformers>=5.0.0",
]

[project.optional-dependencies]
cpu = [
  "torch>=2.9.1",
  "torchvision>=0.24.1",
]
cu128 = [
  "torch>=2.9.1",
  "torchvision>=0.24.1",
]

[project.scripts]
docgemma-mcp = "docgemma.tools.server:main"

[tool.uv]
conflicts = [
  [
    { extra = "cpu" },
    { extra = "cu128" },
  ],
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cpu", extra = "cpu" },
  { index = "pytorch-cu128", extra = "cu128" },
]
torchvision = [
  { index = "pytorch-cpu", extra = "cpu" },
  { index = "pytorch-cu128", extra = "cu128" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true
